[springboot-employee-team][Lebron] refactor: add a exception UnknownEmployeeExcetion handle in EmployeeController

+++ src/main/java/com/thoughtworks/springbootemployee/controller/EmployeeController.java
@@ -2,6 +2,7 @@ package com.thoughtworks.springbootemployee.controller;
 
 import com.thoughtworks.springbootemployee.dto.EmployeeRequestDto;
 import com.thoughtworks.springbootemployee.entity.Employee;
+import com.thoughtworks.springbootemployee.exception.UnknownEmployeeException;
 import com.thoughtworks.springbootemployee.service.EmployeeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -28,6 +29,9 @@ public class EmployeeController {
 
     @GetMapping("/employees/{employeeId}")
     public Optional<Employee> findEmployeeById(@PathVariable int employeeId) {
+        if(!employeeService.findEmployeeById(employeeId).isPresent()){
+            throw new UnknownEmployeeException(employeeId,"Find employee by id failed! Not found!");
+        }
         return employeeService.findEmployeeById(employeeId);
     }
 
@@ -48,11 +52,17 @@ public class EmployeeController {
 
     @PutMapping("/employees/{employeeId}")
     public Employee updateEmployee(@PathVariable int employeeId, @RequestBody EmployeeRequestDto employeeRequestDto) {
+        if(!employeeService.findEmployeeById(employeeId).isPresent()){
+            throw new UnknownEmployeeException(employeeId,"Update employee by id failed! Not found!");
+        }
         return employeeService.updateEmployee(employeeId, employeeRequestDto);
     }
 
     @DeleteMapping("/employees/{employeeId}")
     public void deleteEmployeeById(@PathVariable int employeeId) {
+        if(!employeeService.findEmployeeById(employeeId).isPresent()){
+            throw new UnknownEmployeeException(employeeId,"Delete employee by id failed! Not found!");
+        }
         employeeService.deleteEmployeeById(employeeId);
     }
 }
